+++ src/BlazorMerge/Feature/Merge/MergeService.cs
@@ -3,24 +3,15 @@ using BlazorMerge.Files;
 
 namespace BlazorMerge.Feature.Merge;
 
-public partial class MergeService
+public partial class MergeService(IFileManager fileManager, IMerger merger)
 {
-    private readonly IFileManager _fileManager;
-    private readonly IMerger _merger;
-
-    public MergeService(IFileManager fileManager, IMerger merger)
-    {
-        _fileManager = fileManager;
-        _merger = merger;
-    }
-
     public int MergeEnvironment(MergeOptions options)
     {
-        var mainFileName = Constants.MainFileName;
+        const string mainFileName = Constants.MainFileName;
         var mainFilePath = ConstructPath(options.Path, mainFileName);
-        var readAppSetting = _fileManager.ReadFile(mainFilePath);
+        var readAppSetting = fileManager.ReadFile(mainFilePath);
         var environmentFileName = ReplacePath(options);
-        var readEnvironmentSetting = _fileManager.ReadFile(ConstructPath(options.Path, environmentFileName));
+        var readEnvironmentSetting = fileManager.ReadFile(ConstructPath(options.Path, environmentFileName));
         WriteNewSettingsFile(readAppSetting, readEnvironmentSetting, mainFilePath);
         DeleteSettingsFiles(options);
         WriteNewSettingsFile("{}", "{}", ConstructPath(options.Path, "appsettings.Production.json"));
@@ -30,18 +21,27 @@ public partial class MergeService
 
     private static string ConstructPath(string path, string fileName)
     {
-        var computedPath = path.Trim().EndsWith('\\') || path.Trim().EndsWith('/') ? path : $"{path}\\";
+        var computedPath = path.Trim()
+            .EndsWith('\\') || path.Trim()
+            .EndsWith('/')
+            ? path
+            : $"{path}\\";
         return $"{computedPath}{fileName}";
     }
 
     private void DeleteSettingsFiles(MergeOptions options)
     {
-        var files = _fileManager.ListSettingsFiles(options.Path, OnlySettingsFiles);
+        var files = fileManager.ListSettingsFiles(options.Path, OnlySettingsFiles);
         var settingsFiles = ParseSettingsFiles(files);
         var jsonFilesToDelete = settingsFiles
-            .Where(x => x is {Type: SettingsFileType.Environment, Extension: SettingsFileExtension.Json}).ToList();
-        var brotliFilesToDelete = settingsFiles.Where(x => x is {Extension: SettingsFileExtension.Brotli}).ToList();
-        var gzipFilesToDelete = settingsFiles.Where(x => x is {Extension: SettingsFileExtension.Gzip}).ToList();
+            .Where(x => x is {Type: SettingsFileType.Environment, Extension: SettingsFileExtension.Json})
+            .ToList();
+        var brotliFilesToDelete = settingsFiles.Where(x => x is
+                {Extension: SettingsFileExtension.Brotli})
+            .ToList();
+        var gzipFilesToDelete = settingsFiles.Where(x => x is
+                {Extension: SettingsFileExtension.Gzip})
+            .ToList();
 
         DeleteFiles(jsonFilesToDelete, options);
         DeleteFiles(brotliFilesToDelete, options);
@@ -50,11 +50,12 @@ public partial class MergeService
 
     private static bool OnlySettingsFiles(string fileName)
     {
-        var match = IsAppSettingsFile().Match(fileName);
+        var match = IsAppSettingsFile()
+            .Match(fileName);
         return match.Success;
     }
 
-    private static IList<SettingsFile> ParseSettingsFiles(IEnumerable<string> files)
+    private static List<SettingsFile> ParseSettingsFiles(IEnumerable<string> files)
     {
         var settingsFiles = new List<SettingsFile>();
         foreach (var file in files)
@@ -64,10 +65,7 @@ public partial class MergeService
             settingsFile.FileName = file;
             settingsFile.Type = split[1] switch
             {
-                "Dev" => SettingsFileType.Environment,
-                "Development" => SettingsFileType.Environment,
-                "Production" => SettingsFileType.Environment,
-                "Staging" => SettingsFileType.Environment,
+                "Dev" or "Development" or "Production" or "Staging" => SettingsFileType.Environment,
                 _ => SettingsFileType.Primary
             };
             settingsFile.Extension = split[^1] switch
@@ -87,24 +85,26 @@ public partial class MergeService
     {
         foreach (var file in files.Select(file => ConstructPath(options.Path, file.FileName)))
         {
-            _fileManager.DeleteFile(file);
+            fileManager.DeleteFile(file);
         }
     }
 
     private void WriteNewSettingsFile(string readAppSetting, string readEnvironmentSetting, string mainFileName)
     {
-        var merged = _merger.Merge(readAppSetting, readEnvironmentSetting);
-        _fileManager.WriteFile(mainFileName, merged);
+        var merged = merger.Merge(readAppSetting, readEnvironmentSetting);
+        fileManager.WriteFile(mainFileName, merged);
+        fileManager.WriteGzipFile($"{mainFileName}.gz", merged);
+        fileManager.WriteBrotliFile($"{mainFileName}.br", merged);
     }
 
     private static string ReplacePath(MergeOptions options)
     {
         var environment = options.Environment;
-        var environmentFileNameFormat = Constants.EnvironmentFileName;
-        
+        const string environmentFileNameFormat = Constants.EnvironmentFileName;
+
         return environmentFileNameFormat.Replace("{environment}", environment);
     }
 
-    [GeneratedRegex(@"appsettings(\.\w+)?\.(json|br|gz)$")]
+    [GeneratedRegex(@"appsettings(\.\w+)?\.((json(\.(br|gz))?)|br|gz)$")]
     private static partial Regex IsAppSettingsFile();
 }