+++ src/BlazorMerge/Files/IFileManager.cs
@@ -6,4 +6,6 @@ public interface IFileManager
     void WriteFile(string path, string content);
     void DeleteFile(string path);
     IList<string> ListSettingsFiles(string path, Func<string, bool> parser);
+    void WriteGzipFile(string path, string content);
+    void WriteBrotliFile(string path, string content);
 }