+++ test/BlazorMerge.UnitTests/GlobalUsings.cs
@@ -4,3 +4,5 @@ global using FluentAssertions;
 global using Newtonsoft.Json;
 global using NSubstitute;
 global using NSubstitute.Extensions;
+global using StubLogger;
+global using Xunit;